Laravel

+++ database/migrations/2020_03_24_233813_create_questions_table.php
@@ -18,7 +18,7 @@ class CreateQuestionsTable extends Migration
             $table->timestamps();
             $table->string('title');
             $table->text('body');
-            $table->integer('user_id')->unsigned();
+            $table->unsignedBigInteger('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users');
         });
     }
